Re organize asset group sections into smaller components.
