+++ js/src/components/paid-ads/asset-group/asset-group-editor/asset-field.scss
