+++ js/src/components/paid-ads/asset-group/asset-group-editor/asset-group-editor.scss
@@ -1,4 +1,4 @@
-.gla-asset-group-panel {
+.gla-asset-group-editor {
 	.components-external-link {
 		margin-left: $grid-unit-05;
 	}
@@ -51,11 +51,5 @@
 			padding-left: $grid-unit-20;
 			padding-right: $grid-unit-20;
 		}
-	}
-}
-
-.gla-asset-group-panel-section {
-	.gla-section__body {
-		gap: 0;
 	}
 }
