+++ js/src/components/paid-ads/asset-group/asset-group-editor/asset-group-images-section.js
@@ -1,0 +1,111 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+import { useAdaptiveFormContext } from '~/components/adaptive-form';
+import ImagesSelector from './images-selector';
+import AssetField from './asset-field';
+import Section from '~/components/section';
+import AppDocumentationLink from '~/components/app-documentation-link';
+import { ASSET_IMAGE_SPECS } from '../../assetSpecs';
+
+/**
+ * Renders the images section for an asset group, displaying requirements and allowing users to select and manage images according to
+ * specified asset image specs.
+ *
+ * @param {Object} props - Component props.
+ * @param {Object} props.initialValues - Initial values for the image fields, keyed by asset spec key.
+ * @param {Function} props.refFirstErrorField - Function to bind refs for the first error field.
+ * @param {boolean} props.isSelectedFinalUrl - Indicates if the final URL is selected, enabling/disabling fields.
+ * @param {Function} props.getNumOfIssues - Function to get the number of issues for a given asset key.
+ * @param {Function} props.renderErrors - Function to render error messages for a given asset key.
+ *
+ * @return {JSX.Element} The rendered AssetGroupImagesSection component.
+ */
+const AssetGroupImagesSection = ( {
+	initialValues,
+	refFirstErrorField,
+	isSelectedFinalUrl,
+	getNumOfIssues,
+	renderErrors,
+} ) => {
+	const { values, getInputProps } = useAdaptiveFormContext();
+
+	return (
+		<Section
+			title={ __( 'Images', 'google-listings-and-ads' ) }
+			className="gla-asset-group-section"
+			description={
+				<div className="gla-asset-group-section__primary-description">
+					<p>
+						{ __(
+							'The minimum requirements:',
+							'google-listings-and-ads'
+						) }
+					</p>
+					<ul>
+						<li>
+							{ __(
+								'1x marketing image',
+								'google-listings-and-ads'
+							) }
+						</li>
+						<li>
+							{ __(
+								'1x square marketing image',
+								'google-listings-and-ads'
+							) }
+						</li>
+						<li>{ __( '1x logo', 'google-listings-and-ads' ) }</li>
+					</ul>
+					<p>
+						<AppDocumentationLink
+							context="asset-group"
+							linkId="asset-group-images-learn-more"
+							href="https://support.google.com/google-ads/answer/14530211"
+						>
+							{ __( 'Learn more', 'google-listings-and-ads' ) }
+						</AppDocumentationLink>
+					</p>
+				</div>
+			}
+		>
+			{ ASSET_IMAGE_SPECS.map( ( spec ) => {
+				const initialImageUrls = initialValues[ spec.key ];
+				const imageProps = getInputProps( spec.key );
+
+				return (
+					<AssetField
+						key={ spec.key }
+						ref={ refFirstErrorField.bind( spec.key ) }
+						heading={ spec.heading }
+						subheading={ spec.subheading }
+						help={ spec.help }
+						numOfIssues={ getNumOfIssues( spec.key ) }
+						markOptional={ spec.min === 0 }
+						disabled={ ! isSelectedFinalUrl }
+						initialExpanded={ isSelectedFinalUrl }
+					>
+						<ImagesSelector
+							initialImageUrls={ initialImageUrls }
+							maxNumberOfImages={ spec.getMax( values ) }
+							reachedMaxNumberTip={ spec.getMaxNumberTip(
+								values
+							) }
+							imageConfig={ spec.imageConfig }
+							onChange={ imageProps.onChange }
+						>
+							{ renderErrors( spec.key ) }
+						</ImagesSelector>
+					</AssetField>
+				);
+			} ) }
+		</Section>
+	);
+};
+
+export default AssetGroupImagesSection;
