+++ js/src/components/paid-ads/asset-group/asset-group-editor/images-selector.js
