+++ js/src/components/paid-ads/asset-group/asset-group-editor/index.js
@@ -1,0 +1,1 @@
+export { default } from './asset-group-editor';
