+++ js/src/components/paid-ads/asset-group/asset-group-editor/texts-editor.scss
