+++ js/src/components/paid-ads/asset-group/asset-group-header/assets-loader.scss
