+++ js/src/components/paid-ads/asset-group/asset-group-header/index.js
@@ -1,0 +1,1 @@
+export { default } from './asset-group-header';
