+++ js/src/components/paid-ads/asset-group/asset-group.scss
@@ -33,4 +33,8 @@
 			margin: 0;
 		}
 	}
+
+	.gla-section__body {
+		gap: 0;
+	}
 }
